requirements.txt
Django
psycopg2
Pillow
djangorestframework
markdown
django-filter

mypy
pep8-naming
flake8-builtins
flake8-commas
flake8-quotes
flake8-blind-except
flake8-bugbear
flake8-comprehensions
flake8-pep3101
flake8-docstrings
flake8-coding
flake8-module-name
flake8-pytest
flake8-string-format
flake8-debugger
flake8-print
pytest
pytest-flake8
pytest-isort
pytest-django
pytest-cov
pytest-xdist
pytest-randomly
pycodestyle<2.4.0,>=2.0.0